+++ beaker/data_model/workspace.py
@@ -1,9 +1,8 @@
 from datetime import datetime
-from enum import Enum
 from typing import Dict, List, Optional, Tuple
 
 from .account import Account
-from .base import BaseModel
+from .base import BaseModel, StrEnum
 
 __all__ = [
     "WorkspaceSize",
@@ -64,7 +63,7 @@ class WorkspaceTransferSpec(BaseModel):
     ids: List[str]
 
 
-class Permission(str, Enum):
+class Permission(StrEnum):
     """
     Workspace permission levels.
     """