+++ beaker/services/__init__.py
@@ -10,3 +10,18 @@ from .organization import OrganizationClient
 from .secret import SecretClient
 from .service_client import ServiceClient
 from .workspace import WorkspaceClient
+
+__all__ = [
+    "AccountClient",
+    "ClusterClient",
+    "DatasetClient",
+    "ExperimentClient",
+    "GroupClient",
+    "ImageClient",
+    "JobClient",
+    "NodeClient",
+    "OrganizationClient",
+    "SecretClient",
+    "ServiceClient",
+    "WorkspaceClient",
+]