+++ tests/node_test.py
@@ -2,4 +2,6 @@ from beaker import Beaker
 
 
 def test_node_get(client: Beaker, beaker_node_id: str):
-    assert client.node.get(beaker_node_id).limits.gpu_count > 0
+    gpu_count = client.node.get(beaker_node_id).limits.gpu_count
+    assert gpu_count is not None
+    assert gpu_count > 0